+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.5] - 
+
+- Support for multiple webhooks added
+- Class based code structure added for File Reader and Writer
+- Better Exception Handling
 
 ## [0.0.4] - 02-12-2022